Refactored code a little.
