+++ lib/src/pages/search/search_filters.dart
@@ -1,11 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:template_name/ui/shared/colors/default_colors.dart';
 import 'package:template_name/ui/search/filter_buttons.dart';
-import 'package:template_name/ui/shared/page_resolvers/navigator.dart';
 import 'package:template_name/ui/user_profile/user_button.dart';
 
 class Filters extends StatelessWidget {
-   @override
+  @override
   Widget build(BuildContext context) => SafeArea(
         top: true,
         bottom: true,
@@ -13,40 +12,36 @@ class Filters extends StatelessWidget {
         right: true,
         child: Scaffold(
           backgroundColor: DefaultColors.backgroundColor,
-          body: ListView(
-            shrinkWrap: true,
-            children: <Widget>[
-        Column(crossAxisAlignment: CrossAxisAlignment.start, children: <Widget>[
-           Padding(
-          padding: const EdgeInsets.all(8.0),
-          child: 
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: <Widget>[
-              Row(
+          body: ListView(shrinkWrap: true, children: <Widget>[
+            Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
                 children: <Widget>[
-                  Text(
-                      'Filters',
-                      style: TextStyle(
-                        decoration: TextDecoration.none,
-                        fontSize: 24.0, color: DefaultColors.textColor),
-              ),
-                ],
-              ),
-              exitButton(context),              
-            ],
-            
-          ),
-          
+                  Padding(
+                    padding: const EdgeInsets.all(8.0),
+                    child: Row(
+                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                      children: <Widget>[
+                        Row(
+                          children: <Widget>[
+                            Text(
+                              'Filters',
+                              style: TextStyle(
+                                  decoration: TextDecoration.none,
+                                  fontSize: 24.0,
+                                  color: DefaultColors.textColor),
+                            ),
+                          ],
+                        ),
+                        exitButton(context),
+                      ],
+                    ),
+                  ),
+                  Regions(),
+                ])
+          ]),
         ),
-         Regions(),           
-        ])
-      ]
-      ),
-        ),
-  );
+      );
 }
-
 
 class Regions extends StatelessWidget {
   @override
@@ -55,11 +50,13 @@ class Regions extends StatelessWidget {
       children: <Widget>[
         Padding(
           padding: const EdgeInsets.all(8.0),
-          child:
-        Text('Regions',
-        style: TextStyle(
-          decoration: TextDecoration.none,
-        fontSize: 18.0, color: DefaultColors.textColor),),
+          child: Text(
+            'Regions',
+            style: TextStyle(
+                decoration: TextDecoration.none,
+                fontSize: 18.0,
+                color: DefaultColors.textColor),
+          ),
         ),
         buildFilterButtonRow(['hi', 'hello']),
       ],
